requirements.txt
streamlit
pandas
google-generativeai
Pillow
pymupdf
xlsxwriter
openpyxl